backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
astroid==4.0.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==3.2.2
cachetools==6.2.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
databases==0.9.0
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
google-auth==2.41.1
google-auth-oauthlib==1.2.2
graphviz==0.21
greenlet==3.2.4
griffe==1.13.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
isort==7.0.0
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lxml==6.0.1
mccabe==0.7.0
mcp==1.13.1
oauthlib==3.3.1
openai==1.102.0
openai-agents==0.2.10
passlib==1.7.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.5.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pylint==4.0.2
pypdfium2==4.30.0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.47.3
tiktoken==0.11.0
tomlkit==0.13.3
tqdm==4.67.1
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
